requirements.txt
BCEmbedding==0.1.5
Flask==3.1.0
python-dotenv==1.0.1